feat: Allow arg names (and string arg values) to opt into using the string table
When we did the big re-write, I opted to use inline strings for all args. While this simplifies things, there are many cases where users will use the same arg names for records.

With this change, we still default raw `const char*` strings to use inline strings in the record. However, the user can explicitly construct a `RecordArgumentName` struct now, which allows them to opt-in to using the string table.